Refactoring, remove conv, crud, data/api
