--- src/main/java/com/sfjs/crud/ctrl/PaymentController.java
@@ -1,74 +1,0 @@
-package com.sfjs.crud.ctrl;
-
-import java.util.List;
-
-import org.springframework.graphql.data.method.annotation.Argument;
-import org.springframework.graphql.data.method.annotation.MutationMapping;
-import org.springframework.graphql.data.method.annotation.QueryMapping;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-
-import com.sfjs.crud.svc.PaymentService;
-import com.sfjs.data.api.PaymentData;
-import com.sfjs.data.entity.PaymentEntity;
-
-@RestController
-@EnableWebMvc
-@Transactional
-public class PaymentController extends BaseController<PaymentService, PaymentEntity, PaymentData> {
-
-  @MutationMapping(name = "deletePayment")
-  public Boolean deletePayment(@Argument(name = "id") Long id) {
-    return delete(id);
-  }
-
-  @RequestMapping(path = "/payment/{id}", method = RequestMethod.DELETE)
-  public Boolean delete(@PathVariable(name = "id") Long id) {
-    return super.delete(id);
-  }
-
-  @RequestMapping(path = "/payment/getbyid/{id}", method = RequestMethod.GET)
-  public PaymentData getById(@PathVariable("id") Long id) {
-    return super.getById(id);
-  }
-
-  @RequestMapping(path = "/payment/findbyid/{id}", method = RequestMethod.GET)
-  public PaymentData findById(@PathVariable("id") Long id) {
-    return super.findById(id);
-  }
-
-  @RequestMapping(path = "/payment/findallbyid/{id}", method = RequestMethod.GET)
-  public List<PaymentData> findAllById(@PathVariable("id") Long id) {
-    return super.findAllById(id);
-  }
-
-  @RequestMapping(path = "/payment/findbyname/{name}", method = RequestMethod.GET)
-  public PaymentData findByName(@PathVariable("name") String name) {
-    return super.findByName(name);
-  }
-
-  @RequestMapping(path = "/payment/findallbyname/{name}", method = RequestMethod.GET)
-  public List<PaymentData> findAllByName(@PathVariable("name") String name) {
-    return super.findAllByName(name);
-  }
-
-  @RequestMapping(path = "/payment/findbylabel/{label}", method = RequestMethod.GET)
-  public PaymentData findByLabel(@PathVariable("label") String label) {
-    return super.findByLabel(label);
-  }
-
-  @RequestMapping(path = "/payment/findallbylabel/{label}", method = RequestMethod.GET)
-  public List<PaymentData> findAllByLabel(@PathVariable("label") String label) {
-    return super.findAllByLabel(label);
-  }
-
-  @QueryMapping(name = "findAllPayments")
-  @RequestMapping(path = "/payment/findall", method = RequestMethod.GET)
-  public List<PaymentData> findAll(@Argument(name = "limit") Integer limit) {
-    return super.findAll(limit);
-  }
-}
