--- src/main/java/com/sfjs/data/api/BusinessData.java
@@ -1,13 +1,0 @@
-package com.sfjs.data.api;
-
-import com.sfjs.data.core.Business;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class BusinessData extends Business {
-
-  @Getter
-  @Setter
-  private AccountData account;
-}
