--- src/main/java/com/sfjs/data/api/PayDetailsData.java
@@ -1,11 +1,0 @@
-package com.sfjs.data.api;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class PayDetailsData {
-
-  @Getter @Setter private float payScaleMin; // : Float
-  @Getter @Setter private float payScaleMax; // : Float
-  @Getter @Setter private String payOption; // : String
-}
