--- src/main/java/com/sfjs/data/api/PaymentData.java
@@ -1,27 +1,0 @@
-package com.sfjs.data.api;
-
-import com.sfjs.data.core.Payment;
-
-import jakarta.persistence.Transient;
-import lombok.Getter;
-import lombok.Setter;
-
-public class PaymentData extends Payment {
-
-  @Getter
-  @Setter
-  @Transient
-  private String checkoutToken; // from helcim - don't store
-
-  @Getter
-  @Setter
-  private BusinessData business;
-
-  @Getter
-  @Setter
-  private FellowData fellow;
-
-  @Getter
-  @Setter
-  private String secretToken; // from helcim - store encrypted
-}
