--- src/main/java/com/sfjs/data/api/RoleData.java
@@ -1,7 +1,0 @@
-package com.sfjs.data.api;
-
-import com.sfjs.data.core.Role;
-
-public class RoleData extends Role {
-
-}
