+++ src/main/java/com/sfjs/data/entity/ProfileElementData.java
@@ -1,8 +1,7 @@
-package com.sfjs.data.api;
+package com.sfjs.data.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.sfjs.data.BaseObject;
-import com.sfjs.data.entity.FellowProfileEntity;
 
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
