+++ src/main/java/com/sfjs/gql/svc/JobListingService.java
@@ -15,13 +15,13 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.sfjs.data.entity.AccountEntity;
 import com.sfjs.data.entity.BusinessEntity;
 import com.sfjs.data.entity.InterviewProcessEntity;
+import com.sfjs.data.entity.JobListingElementData;
 import com.sfjs.data.entity.JobListingEntity;
 import com.sfjs.jpa.repo.BaseRepository;
 import com.sfjs.jpa.repo.InterviewProcessRepository;
 import com.sfjs.jpa.repo.JobListingRepository;
 import com.sfjs.data.BaseObject;
 import com.sfjs.data.api.JobListingData;
-import com.sfjs.data.api.JobListingElementData;
 import com.sfjs.data.api.PayDetailsData;
 import com.sfjs.data.core.HybridDetails;
 import com.sfjs.data.core.InterviewProcess;
