+++ src/main/java/com/sfjs/gql/svc/SignupService.java
@@ -26,6 +26,7 @@ import com.sfjs.data.entity.FellowProfileEntity;
 import com.sfjs.data.entity.FellowEntity;
 import com.sfjs.data.entity.HobbyEntity;
 import com.sfjs.data.entity.LinkEntity;
+import com.sfjs.data.entity.ProfileElementData;
 import com.sfjs.data.entity.RoleEntity;
 import com.sfjs.jpa.repo.AccomplishmentRepository;
 import com.sfjs.jpa.repo.AccountRepository;
@@ -44,7 +45,6 @@ import com.sfjs.jpa.repo.ProfileRepository;
 import com.sfjs.jpa.repo.RoleRepository;
 import com.sfjs.data.BaseObject;
 import com.sfjs.data.api.FellowProfileData;
-import com.sfjs.data.api.ProfileElementData;
 import com.sfjs.data.core.Accomplishment;
 import com.sfjs.data.core.Award;
 import com.sfjs.data.core.BookOrQuote;
